backend/requirements.txt
fastapi
uvicorn
pyserial
